o zi neproductiva

+++ server/src/api/controllers/User.js
@@ -124,8 +124,12 @@ async function InsertNewMessage(req, res) {
     roomID,
     messageBody
   );
+  if (result === "FAILED") {
+    console.log("Error storage message!");
+    return handleResponse(req, res, 412, " DataBase Error ");
+  }
 
-  return handleResponse(req, res, 200, { result });
+  return handleResponse(req, res, 200, { InserNewMessage: "SUCCES" });
 }
 
 // creata a new grup
@@ -145,20 +149,27 @@ async function CreateNewRoom(req, res) {
     return handleResponse(req, res, 410, "Invalid Request Parameters ");
   }
 
-  var roomResult = await InsertNewRoomData(RoomName, Private, uuidRoom);
+  const roomResult = await InsertNewRoomData(RoomName, Private, uuidRoom);
 
-  if (userSearchListID !== null && userID !== null) {
-    await InsertParticipantData(uuidRoom, userSearchListID);
-    await InsertParticipantData(uuidRoom, userID);
+  if (roomResult === "FAILED") {
+    return handleResponse(req, res, 412, " DataBase Error ");
   }
 
-  // TO DO : de trimis rezultat pozitiv
+  if (userSearchListID !== null && userID !== null) {
+    var partResult = await InsertParticipantData(uuidRoom, userSearchListID);
+    if (partResult === "FAILED")
+      return handleResponse(req, res, 412, " DataBase Error ");
+    partResult = await InsertParticipantData(uuidRoom, userID);
+    if (partResult === "FAILED")
+      return handleResponse(req, res, 412, " DataBase Error ");
+  }
+
   return handleResponse(req, res, 200, {
-    "Create New Private Conversation - SUCCES": "ok",
+    CreateNewPrivateConversation: "SUCCES",
   });
 }
 
-// delete room
+// delete group or private conversation (room)
 async function DeleteRoom(req, res) {
   const roomID = req.body.roomID;
 
@@ -167,10 +178,24 @@ async function DeleteRoom(req, res) {
   }
 
   var res_deletemes = await DeleteAllMessageFromRoom(roomID);
-  var res_deletepart = await DeleteAllParticipantsFromRoom(roomID);
-  var res_deleteroom = await DeleteRoomData(roomID);
+  if (res_deletemes === "FAILED") {
+    console.log("FAILED to delete messages from room!");
+    return handleResponse(req, res, 412, " DataBase Error ");
+  }
 
-  return handleResponse(req, res, 200, { res_deleteroom });
+  var res_deletepart = await DeleteAllParticipantsFromRoom(roomID);
+  if (res_deletepart === "FAILED") {
+    console.log("FAILED to delete participants from room!");
+    return handleResponse(req, res, 412, " DataBase Error ");
+  }
+
+  var res_deleteroom = await DeleteRoomData(roomID);
+  if (res_deleteroom === "FAILED") {
+    console.log("FAILED to delete room!");
+    return handleResponse(req, res, 412, " DataBase Error ");
+  }
+
+  return handleResponse(req, res, 200, { DeleteRoom: "SUCCESS" });
 }
 
 // Create a new room (group)
@@ -191,10 +216,18 @@ async function CreateNewRoom_Group(req, res) {
 
   // adaugare camera noua in tabela
   var roomResult = await InsertNewRoomData(NewGroupName, Type, uuidRoom);
+  if (roomResult === "FAILED") {
+    console.log("FAILED - insert new room! ");
+    return handleResponse(req, res, 412, " DataBase Error ");
+  }
 
   // adaugare participant la camera creata mai sus
   if (userID !== null) {
-    await InsertParticipantData(uuidRoom, userID);
+    var partResult = await InsertParticipantData(uuidRoom, userID);
+    if (partResult === "FAILED") {
+      console.log("FAILED - inser new participant in room! ");
+      return handleResponse(req, res, 412, " DataBase Error ");
+    }
   }
 
   return handleResponse(req, res, 200, "New Group Created successful");
@@ -211,9 +244,16 @@ async function AddNewMemberInGroup(req, res) {
 
   var participantDetails = await GetParticipantByID(userSearchListID, roomID);
 
-  if (!participantDetails.length) {
+  if (participantDetails === "FAILED") {
+    console.log("FAILED - get participant! ");
+    return handleResponse(req, res, 412, " DataBase Error ");
+  } else if (!participantDetails.length) {
     var result = await AddNewMemberInGroupData(roomID, userSearchListID);
-    return handleResponse(req, res, 200, { result });
+    if (result === "FAILED") {
+      console.log("FAILED - add participant to group! ");
+      return handleResponse(req, res, 412, " DataBase Error ");
+    }
+    return handleResponse(req, res, 200, { AddParticipant: "SUCCES" });
   } else {
     return handleResponse(req, res, 200, "User is Already a member");
   }
@@ -228,6 +268,11 @@ async function GetPartList(req, res) {
   }
 
   var participantsRoomList = await GetPartListData(roomID);
+  if (participantsRoomList === "FAILED") {
+    console.log("FAILED - get participants list! ");
+    return handleResponse(req, res, 412, " DataBase Error ");
+  }
+
   return handleResponse(req, res, 200, { participantsRoomList });
 }
 
@@ -253,17 +298,20 @@ async function UpdateProfilePicture(req, res) {
     return handleResponse(req, res, 410, "Invalid Request Parameters ");
   }
 
-  const WriteResult = await WriteFileToDisc(path, fileName, NewPicture);
+  const WriteFileResult = await WriteFileToDisc(path, fileName, NewPicture);
 
-  if (WriteResult === "SUCCESS") {
-    //update database path
-    const result = UpdateAvatarPathData(userID, path + fileName);
-    if (result === "FAILED") {
-      return handleResponse(req, res, 412, " DataBase Error ");
-    }
+  if (WriteFileResult === "FAILED") {
+    console.log("FAILED - Write File To Disc");
+    return handleResponse(req, res, 413, " Write File Error ");
+  }
+  //update database avatar's path
+  const result = UpdateAvatarPathData(userID, path + fileName);
+  if (result === "FAILED") {
+    console.log("FAILED - update user avatar path!");
+    return handleResponse(req, res, 412, " DataBase Error ");
   }
 
-  return handleResponse(req, res, 200, " Execution Succed ");
+  return handleResponse(req, res, 200, { UpdateProfilePicture: "SUCCESS" });
 }
 
 module.exports = {
